// src/app/services/models/contract-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { DocumentShortDto } from '../models/document-short-dto';
import { EmployeeShortDto } from '../models/employee-short-dto';
export interface ContractDto {
  documents?: Array<DocumentShortDto>;
  employee?: EmployeeShortDto;
  endDate?: string;
  id?: number;
  salary?: number;
  startDate?: string;
  status?: 'DRAFT' | 'PENDING' | 'ACTIVE' | 'EXPIRED' | 'TERMINATED' | 'RENEWED';
  terminationReason?: string;
  type?: 'CDI' | 'CDD' | 'STAGE' | 'INTERIM' | 'FREELANCE' | 'APPRENTICESHIP' | 'PART_TIME';
  workHoursPerWeek?: number;
}