// src/app/services/models/contract-create-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ContractCreateDto {
  employeeId?: number;
  endDate?: string;
  salary?: number;
  startDate?: string;
  status?: 'DRAFT' | 'PENDING' | 'ACTIVE' | 'EXPIRED' | 'TERMINATED' | 'RENEWED';
  type?: 'CDI' | 'CDD' | 'STAGE' | 'INTERIM' | 'FREELANCE' | 'APPRENTICESHIP' | 'PART_TIME';
  workHoursPerWeek?: number;
}